Print daddr_t's with %j and intmax_t.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@106673 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/ufs/ufs/ufs_extattr.c
@@ -52,6 +52,7 @@
 #include <sys/lock.h>
 #include <sys/dirent.h>
 #include <sys/extattr.h>
+#include <sys/stdint.h>
 #include <sys/sysctl.h>
 
 #include <vm/uma.h>
@@ -911,8 +912,8 @@ ufs_extattr_get(struct vnode *vp, int attrnamespace, const char *name,
 		 * is to coerce this to undefined, and let it get cleaned
 		 * up by the next write or extattrctl clean.
 		 */
-		printf("ufs_extattr_get (%s): inode number inconsistency (%d, %lld)\n",
-		    mp->mnt_stat.f_mntonname, ueh.ueh_i_gen, ip->i_gen);
+		printf("ufs_extattr_get (%s): inode number inconsistency (%d, %jd)\n",
+		    mp->mnt_stat.f_mntonname, ueh.ueh_i_gen, (intmax_t)ip->i_gen);
 		error = ENOATTR;
 		goto vopunlock_exit;
 	}
@@ -1192,8 +1193,8 @@ ufs_extattr_rm(struct vnode *vp, int attrnamespace, const char *name,
 		 * coerce this to undefined, and let it get cleaned up by
 		 * the next write or extattrctl clean.
 		 */
-		printf("ufs_extattr_rm (%s): inode number inconsistency (%d, %lld)\n",
-		    mp->mnt_stat.f_mntonname, ueh.ueh_i_gen, ip->i_gen);
+		printf("ufs_extattr_rm (%s): inode number inconsistency (%d, %jd)\n",
+		    mp->mnt_stat.f_mntonname, ueh.ueh_i_gen, (intmax_t)ip->i_gen);
 		error = ENOATTR;
 		goto vopunlock_exit;
 	}
